// api/seller/client/models/StoreUser.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { StoreRole } from './StoreRole';

/**
 * Model StoreUser
 */
export type StoreUser = {
  updatedAt: string;
  createdAt: string;
  isOwner: boolean | null;
  role: StoreRole;
  userId: string;
  storeId: string;
  id: number;
};